update pidcontroller continuous call, remove old command-based template
