+++ wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/templates/commandbased/Constants.java
@@ -5,7 +5,7 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package edu.wpi.first.wpilibj.templates.commandbasednew;
+package edu.wpi.first.wpilibj.templates.commandbased;
 
 /**
  * The Constants class provides a convenient place for teams to hold robot-wide numerical or boolean
