--- wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/templates/commandbasednew/commands/ExampleCommand.java
@@ -1,29 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2018-2019 FIRST. All Rights Reserved.                        */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package edu.wpi.first.wpilibj.templates.commandbasednew.commands;
-
-import edu.wpi.first.wpilibj.templates.commandbasednew.subsystems.ExampleSubsystem;
-import edu.wpi.first.wpilibj2.command.SendableCommandBase;
-
-/**
- * An example command that uses an example subsystem.
- */
-public class ExampleCommand extends SendableCommandBase {
-  @SuppressWarnings({"PMD.UnusedPrivateField", "PMD.SingularField"})
-  private final ExampleSubsystem m_subsystem;
-
-  /**
-   * Creates a new ExampleCommand.
-   *
-   * @param subsystem The subsystem used by this command.
-   */
-  public ExampleCommand(ExampleSubsystem subsystem) {
-    m_subsystem = subsystem;
-    addRequirements(subsystem);
-  }
-}
