--- wpilibjExamples/src/main/java/edu/wpi/first/wpilibj/templates/commandbasednew/subsystems/ExampleSubsystem.java
@@ -1,27 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2018-2019 FIRST. All Rights Reserved.                        */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package edu.wpi.first.wpilibj.templates.commandbasednew.subsystems;
-
-import edu.wpi.first.wpilibj2.command.SendableSubsystemBase;
-
-public class ExampleSubsystem extends SendableSubsystemBase {
-  /**
-   * Creates a new ExampleSubsystem.
-   */
-  public ExampleSubsystem() {
-
-  }
-
-  /**
-   * Will be called periodically whenever the CommandScheduler runs.
-   */
-  @Override
-  public void periodic() {
-
-  }
-}
